[FTR] Movement of SelectActivity RecyclerView Adapter
Feature Description

Movement of the recyclerview adapter for the selectactivity activity to its
own file to address code cleanup in preparation of supporting the query operations
with Firebase database.

Related Issues
- #89
